+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ## `@svizzle/barchart` v0.5.0 (next)
 
-- add prop `titleFontSize`
+- add prop `refs`
 - add prop `selectedKeys`
+- add prop `titleFontSize`
 - don't show the axis with positive and zeroes or negative and zeroes
 
 ## `@svizzle/choropleth` v0.4.0 (next)