+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Svizzle changelog
 
-## next
+## 20200908
 
-## `@svizzle/choropleth` (next)
+## `@svizzle/choropleth` v0.5.0
 
 - document `topoToGeo`
 - add prop `focusedKey`